refactor, rebased, redone, REEE
