--- src/sql/SQLAlchemy/utils.py
@@ -1,75 +1,0 @@
-import os
-from sqlalchemy.engine.url import URL
-from sqlalchemy.orm import sessionmaker
-from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.ext.compiler import compiles
-from sqlalchemy.sql import expression
-from sqlalchemy import create_engine, inspect, MetaData, Table, Column, Integer, String, DateTime
-
-# Global variables / RDS_HOSTNAME is configurable from Elastic Beanstalk.
-if 'RDS_HOSTNAME' in os.environ:
-	postgres_db = {
-		'drivername':'postgres',
-		'username': os.environ['RDS_USERNAME'],
-		'password': os.environ['RDS_PASSWORD'],
-		'host': os.environ['RDS_HOST'],
-		'port': os.environ['RDS_PORT']
-		}
-else:
-	postgres_db = {
-		'drivername':'postgres',
-		'username': os.environ['PG_USERNAME'],
-		'password': os.environ['PG_PASSWORD'],
-		'host': os.environ['PG_HOST'],
-		'port': os.environ['PG_PORT']
-		}
-
-DB_URI = URL(**postgres_db)
-ENGINE = create_engine(DB_URI)
-Base = declarative_base()
-
-def inspect_db(name,ENGINE):
-	'''
-	Inspect database to check whether or not a table exists.
-	:name: String , name of table
-	'''
-	inspector = inspect(ENGINE)
-
-	if name in inspector.get_table_names():
-		return True
-
-# For PostgreSQL UTC timestamp custom function
-class utcnow(expression.FunctionElement):
-	type = DateTime()
-
-@compiles(utcnow, 'postgresql')
-def pg_utcnow(element,compiler,**kw):
-	return "TIMEZONE('utc', CURRENT_TIMESTAMP)"
-
-class CreateTable(object):
-	'''
-	CreateTable : no default name, must inherit Base object and specify a name
-	with __tablename__.
-	inheritable object that can instantiate different tweets tables
-	if we need to a) create new tables with different names,
-	or automate table creation in the future.
-	'''
-	id = Column(Integer, primary_key=True)
-	tweet_id = Column(String)
-	user_id = Column(String)
-	username = Column(String)
-	screenname = Column(String)
-	link_to_profile = Column(String, key='profile_link')
-	permalink = Column(String, key='link')
-	language = Column(String)
-	time = Column(DateTime)
-	timestamp = Column(DateTime(timezone=True), server_default=pg_utcnow())
-	retweets = Column(Integer)
-	likes = Column(Integer)
-	text = Column(String)
-
-class CreateTweetsTable(CreateTable, Base):
-	__tablename__ = 'tweets'
-
-
